backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
dataclasses-json==0.6.7
deprecated==1.2.18
distro==1.9.0
docx2txt==0.8
durationpy==0.9
emoji==2.14.1
eval-type-backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.38.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.5.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.1
langchain==0.3.20
langchain-chroma==0.2.2
langchain-community==0.3.19
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.13
lxml==5.3.1
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.21.0
openai==1.65.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
posthog==3.19.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pygments==2.19.1
pypdf==5.3.1
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pyyaml==6.0.2
rapidfuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.38
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
unstructured==0.16.25
unstructured-client==0.31.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
